Only buffer at logical boundaries
It is all too easy to produce an invalid file if we allow
bypass writes in combination with buffering. So at a small-ish
cost in terms of buffering efficiency (we will be less efficient
with small files inside the ZIP) flush at every "regular" segment
boundary, such as a local header, data descriptor or the cdir. This
way we can make the usage a bit safer and less annoying for the
end-consumer.

A larger buffer could be introduced in the output enumerator though.